bug fixes and improvements

+++ renderer/include/Utils/FileIO.h
@@ -21,8 +21,8 @@ public:
 
     static std::ifstream::pos_type getFileSize(const std::string& filename);
 
-    static std::string loadFromTextFile(const std::string& filename, uint* sizePtr = nullptr);
-    static std::vector<char> loadFromBinaryFile(const std::string& filename, uint* sizePtr = nullptr);
+    static std::string loadFromTextFile(const std::string& filename, size_t* sizePtr = nullptr);
+    static std::vector<char> loadFromBinaryFile(const std::string& filename, size_t* sizePtr = nullptr);
     static unsigned char* loadImage(const std::string& filename, int* width, int* height, int* channels, int desiredChannels = 0);
     static unsigned char* loadImageFromMemory(const unsigned char* data, int size, int* width, int* height, int* channels, int desiredChannels = 0);
     static float* loadImageHDR(const std::string& filename, int* width, int* height, int* channels, int desiredChannels = 0);
